configuration de la sécurité avec restriction de route en fct des roles, mise en place du forum avec la generation des tables et premiere requete findAll et find by id"

+++ src/main/java/com/api/mushroom/service/EdibilityService.java
@@ -81,8 +81,12 @@ public class EdibilityService {
 
 
     // delete : Supprimer un enregistrement
-    public void delete(Long id) {
-        edibilityJpaRepository.deleteById(id);
+    public boolean delete(Long id) {
+        if(edibilityJpaRepository.existsById(id)){
+            edibilityJpaRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
